extract sft entrypoint
